backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djongo==1.3.6
pymongo==3.12.3
python-decouple==3.8
django-cors-headers==4.3.0
Pillow==10.1.0

# ML Libraries
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2

# Data Processing
openpyxl==3.1.2
xlrd==2.0.1

# PDF Generation
reportlab==4.0.7
PyPDF2==3.0.1

# Utilities
python-dateutil==2.8.2
pytz==2023.3
